Card converted to Typed Functional react component
